moved login logic
